fix(relay): increase page size for metrics to 4096 (#8646)
The default here is 2 which is nowhere near enough of a batch-size for
us to read all perf events generated by the kernel when it is actually
relaying data via eBPF (we generate 1 perf event per relayed packet). If
we don't read them fast enough, the kernel has to drop some, meaning we
skew our metrics as to how much data we've relayed via eBPF.

This has been tested in my local setup and I've seen north of 500 events
being read in a single batch now.

---------

Signed-off-by: Thomas Eizinger <thomas@eizinger.io>