chore(ci): Upload signed artifacts to run on workflow_dispatch (#8778)
When testing certain PRs, it's helpful to have signed release builds for
various platforms.

These can be built by manually triggering their respective workflow from
the GitHub UI. In these cases, we want to upload the artifacts to the
workflow run, but _not_ upload the artifacts to the release.

We only want to upload artifacts to the release if the `github.ref_name`
is `main`.